==> R/04-analysis/01-settings.R <==
# Notes ----------------------------------------------------------------------------------
#   Goal:   Settings for prediction models
#   Time:   None


# Set values -----------------------------------------------------------------------------
  # Outcome variable
  outcome_var = 'dbwt'
  # outcome_var = 'gestation'
